refactor: move evm executor to own crate
