+++ Cargo.toml
@@ -7,7 +7,8 @@ members = [
     "cli/test-utils",
     "config",
     "fmt",
-    "ui"
+    "ui",
+    "evm"
 ]
 
 # Binary size optimizations
