+++ evm/Cargo.toml
@@ -1,0 +1,39 @@
+[package]
+name = "foundry-evm"
+version = "0.1.0"
+edition = "2021"
+
+# TODO: We can probably reduce dependencies here or in the forge crate
+[dependencies]
+foundry-utils = { path = "./../utils" }
+
+# Encoding/decoding
+serde_json = "1.0.67"
+serde = "1.0.130"
+hex = "0.4.3"
+ethers = { git = "https://github.com/gakonst/ethers-rs", default-features = false, features = ["solc-full"] }
+
+# Error handling
+eyre = "0.6.5"
+thiserror = "1.0.29"
+
+# Logging
+tracing = "0.1.26"
+tracing-subscriber = "0.2.20"
+
+# Threading/futures
+tokio = { version = "1.10.1" }
+parking_lot = "0.12.0"
+futures = "0.3.21"
+once_cell = "1.9.0"
+
+# EVM
+bytes = "1.1.0"
+hashbrown = "0.12"
+revm = { package = "revm", git = "https://github.com/bluealloy/revm", default-features = false, features = ["std", "k256"] }
+
+# Fuzzer
+proptest = "1.0.0"
+
+# Display
+ansi_term = "0.12.1"
