+++ evm/src/executor/inspector/stack.rs
