+++ evm/src/lib.rs
@@ -1,0 +1,19 @@
+/// Decoding helpers
+pub mod decode;
+
+/// Call trace arena, decoding and formatting
+pub mod trace;
+
+/// Debugger arena
+pub mod debugger;
+
+/// Forge test execution backends
+pub mod executor;
+pub use executor::abi;
+
+// Re-exports
+pub use ethers::types::Address;
+pub use hashbrown::HashMap;
+
+use once_cell::sync::Lazy;
+pub static CALLER: Lazy<Address> = Lazy::new(Address::random);
