+++ forge/Cargo.toml
@@ -6,6 +6,7 @@ license = "MIT OR Apache-2.0"
 
 [dependencies]
 foundry-utils = { path = "./../utils" }
+foundry-evm = { path = "./../evm" }
 
 ethers = { git = "https://github.com/gakonst/ethers-rs", default-features = false, features = ["solc-full"] }
 eyre = "0.6.5"
@@ -22,15 +23,7 @@ tracing-subscriber = "0.2.20"
 proptest = "1.0.0"
 rayon = "1.5"
 rlp = "0.5.1"
-
-bytes = "1.1.0"
-thiserror = "1.0.29"
-revm = { package = "revm", git = "https://github.com/bluealloy/revm", default-features = false, features = ["std", "k256"] }
-hashbrown = "0.12"
 once_cell = "1.9.0"
-parking_lot = "0.12.0"
-futures = "0.3.21"
-ansi_term = "0.12.1"
 comfy-table = "5.0.0"
 
 [dev-dependencies]
